Implement unit test demonstrating svn url path problem
Make computeChangePaths in SvnScmProvider package private to be able to test it directly with a unit test. Implement a test demonstrating the issue that occurs, when a svn repository is located on the root level of an url (e.g. http://svnserver/). SVNKit then returns "" as root repository path instead of "/" and that causes a crash - demonstrated by the test case implemented in this commit.

See internal #SUPPORT-14039 jira ticket.